Refactor DailyCalendar instantiation in ImplementationTests
The Microsoft.Extensions.Logging import has been removed from ImplementationTests.cs as it was not used. In addition, the construction of a DailyCalendar object has been updated to use DateTime objects instead of Ticks. Furthermore, the method call to trigger.GetNextFireTimeUtc() has been replaced with the variable nextFireTime for consistency.