Add ControllingWatcher helper class and update Notify signature
A new ControllingWatcher class inside Helpers has been added to the Quartz.Impl.UnitTests. It is designed to help with scheduler execution control during tests. Meanwhile, the signature of Notify method in IDebugWatcher interface and its usage was updated to only take two parameters instead of three. The ControllingWatcher class previously in ClusteredSchedulerTests was removed as a result of this refactoring.