Add 'Completed' SchedulerExecutionStep and PersistentJob
Added a 'Completed' step to the 'SchedulerExecutionStep' enumeration and the 'ControllingWatcher' class. Created a new 'PersistentJob' class for unit tests. Updated 'DeleteJobIfSingleReferenceAsync' method in the 'RavenJobStore' to check if the job is durable before deleting. The 'NotifyDebugWatcher' method now takes the 'SchedulerExecutionStep.Completed' as an argument. The 'SingleSchedulerTests' class now includes a test for persistent job data update during completion.