Add self-replacing and terminating jobs, fix blocking issue in job rescheduling
New job types, SelfReplacingJob and TerminatingJob, are added for unit tests. An issue in RavenJobStore where a non-concurrent job when rescheduling itself caused trigger blocking has been resolved. This change also contains improvements for job handling when a job is blocked, ensuring non-related triggers are not affected. The project version is updated to 0.8.0-prerelease.