Convert WaitForIndexing to async and update version to 1.0.0-beta.2
The method WaitForIndexing() has been changed to an asynchronous method WaitForIndexingAsync() that takes an array of names, checks if any index is in an Error state, and has a delay between checks. The method is now called with the name of the index. Also, the Quartz.Impl.RavenJobStore version has been updated from 1.0.0-beta.1 to 1.0.0-beta.2.